pyproject.toml
[project]
name = "representations-across-sizes"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Alexander Loftus", email = "alexloftus2004@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "beartype>=0.19.0",
    "dadapy",
    "datasets>=3.1.0",
    "ipykernel>=6.29.5",
    "ipython>=8.29.0",
    "ipywidgets>=8.1.5",
    "jaxtyping>=0.2.36",
    "lovely-tensors>=0.1.17",
    "nnsight>=0.3.6",
    "numba>=0.60.0",
    "pytest>=8.3.3",
    "rich[jupyter]>=13.9.4",
    "scikit-dimension>=0.3.4",
    "scipy>=1.14.1",
    "seaborn>=0.13.2",
    "torch>=2.5.1",
    "torchtyping>=0.1.5",
    "tqdm>=4.66.6",
    "transformers>=4.46.2",
]

[project.optional-dependencies]
ruff = [
    "ipykernel>=6.29.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "python-dotenv>=1.0.1",
]

[tool.uv.sources]
dadapy = { git = "https://github.com/sissa-data-science/DADApy" }
